+++ tests/testthat/test-utils.R
@@ -252,3 +252,14 @@ testthat::test_that("defunction recursively goes down a list", {
     y
   )
 })
+
+testthat::test_that("create_renv_lockfile creates a lock file during the execution", {
+  old_plan <- future::plan(future::sequential)
+  withr::defer(future::plan(old_plan))
+
+  renv_file_name <- "teal_app.lock"
+  withr::defer(file.remove(renv_file_name))
+  promise <- create_renv_lockfile(TRUE, renv_file_name)
+
+  testthat::expect_true(file.exists(renv_file_name))
+})