+++ NEWS.md
@@ -3,6 +3,10 @@
 ### Enhancement
 * Provided progress bar for modules loading and data filtering during teal app startup.
 
+### New features
+
+* Possibility to download lockfile to restore app session for reproducibility.
+
 ### Miscellaneous
 * Filter mapping display is no longer coupled to the snapshot manager.